# Conflicts:
#   jpf-market-api/src/main/resources/config/config.properties

+++ jpf-api/src/main/java/com/joiest/jpf/dto/GetShopBargainRequestRequest.java
@@ -1,5 +1,6 @@
 package com.joiest.jpf.dto;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class GetShopBargainRequestRequest {
@@ -21,7 +22,7 @@ public class GetShopBargainRequestRequest {
     /**
      * 最低豆数
      */
-    private Integer minDou;
+    private BigDecimal minDou;
 
     /**
      * 是否开启此收豆信息 0=不开启 1=开启
@@ -78,11 +79,11 @@ public class GetShopBargainRequestRequest {
         this.offRate = offRate;
     }
 
-    public Integer getMinDou() {
+    public BigDecimal getMinDou() {
         return minDou;
     }
 
-    public void setMinDou(Integer minDou) {
+    public void setMinDou(BigDecimal minDou) {
         this.minDou = minDou;
     }
