+++ jpf-domain/src/main/java/com/joiest/jpf/common/custom/PayShopBargainRequestCustom.java
@@ -1,6 +1,7 @@
 package com.joiest.jpf.common.custom;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class PayShopBargainRequestCustom implements Serializable {
@@ -22,7 +23,7 @@ public class PayShopBargainRequestCustom implements Serializable {
     /**
      * 最低豆数
      */
-    private Integer minDou;
+    private BigDecimal minDou;
 
     /**
      * 是否开启此收豆信息 0=不开启 1=开启
@@ -67,11 +68,11 @@ public class PayShopBargainRequestCustom implements Serializable {
         this.offRate = offRate;
     }
 
-    public Integer getMinDou() {
+    public BigDecimal getMinDou() {
         return minDou;
     }
 
-    public void setMinDou(Integer minDou) {
+    public void setMinDou(BigDecimal minDou) {
         this.minDou = minDou;
     }
