+++ jpf-domain/src/main/java/com/joiest/jpf/common/custom/PayShopProductCustomer.java
@@ -38,7 +38,7 @@ public class PayShopProductCustomer implements Serializable {
     /**
      * 豆
      */
-    private Integer dou;
+    private BigDecimal dou;
 
     /**
      * 产品简介
@@ -284,11 +284,11 @@ public class PayShopProductCustomer implements Serializable {
         this.bid = bid;
     }
 
-    public Integer getDou() {
+    public BigDecimal getDou() {
         return dou;
     }
 
-    public void setDou(Integer dou) {
+    public void setDou(BigDecimal dou) {
         this.dou = dou;
     }
